+++ prototype/io.js
@@ -7,6 +7,16 @@ var dataDir = process.env.PLACEHOLDER_DATA || path.join( __dirname, '../data/');
 var graphPath = path.join( dataDir, 'graph.json' );
 var storePath = path.join( dataDir, 'store.sqlite3' );
 
+// WIP
+var Database = require('../wip/Database');
+var db = new Database('./db' );
+var tokenize = require('../wip/test_tokenize').tokenize.bind({
+  graph: {
+    hasToken: db.hasSubject.bind( db ),
+    hasTokenAutocomplete: db.hasSubjectAutocomplete.bind( db )
+  }
+});
+
 // load data from disk
 module.exports.load = function( opts ){
   this.store.open( storePath );
@@ -17,6 +27,12 @@ module.exports.load = function( opts ){
     this.graph.nodes = graph.nodes;
     this.graph.edges = graph.edges;
   }
+
+  // WIP
+  this.wip = {
+    db: db,
+    tokenize: tokenize
+  };
 };
 
 // save data to disk
