+++ server/routes/tokenize.js
@@ -4,7 +4,10 @@ module.exports = function( req, res ){
   // placeholder
   var ph = req.app.locals.ph;
 
-  var tokens = ph.tokenize( req.query.text );
+  // var tokens = ph.tokenize( req.query.text );
 
-  res.status(200).json( tokens );
+  ph.wip.tokenize( req.query.text, ( err, groups ) => {
+    res.status(200).json( groups );
+  });
+
 };
